+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All significant changes to this project are recorded in this file. Releases foll
 - Functionality to automatically set opacity values
 - Images can now be manipulated through modifiers
 - Added a few basic modifier functions
+- Added animations and visuals for them
+
+### Changed
+- Image will now infer pixel count from other arguments if it is not stated
 
 ## [0.1.0] - 2025-07-09